requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
numpy
datasets
accelerate
chromadb
pydantic
python-multipart
huggingface-hub
aiofiles